+++ package.json
@@ -87,7 +87,6 @@
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-typescript": "^0.14.0",
     "file-loader": "^6.2.0",
-    "graphql-tag": "^2.12.6",
     "history": "^4.9.0",
     "html-webpack-plugin": "^5.5.0",
     "https-browserify": "^1.0.0",
@@ -182,7 +181,6 @@
     "@types/react-router-dom": "^5.3.3",
     "@uniswap/sdk": "^3.0.3",
     "@xenova/transformers": "^2.17.0",
-    "apollo-boost": "^0.4.7",
     "bech32": "^1.1.3",
     "big.js": "^5.2.2",
     "bignumber.js": "^9.0.0",
